requirements.txt
bilibili-api-python>=17.1.4
jwcrypto

fastapi
uvicorn

sqlalchemy
aiosqlite
cachetools

ruamel.yaml
pydantic
pydantic-core

httpx
